% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySignature.R
\name{applySignature}
\alias{applySignature}
\title{Title applySignature}
\usage{
applySignature(
  matrix,
  geneSymbols,
  signature = c("Gempred", "Puleo", "Purist", "McpCount"),
  norm = "none",
  scale = "none"
)
}
\arguments{
\item{geneSymbols}{gene symbols, a vector of same length as the number of rows in newex}

\item{signature}{signature to apply, can be Gempred, Puleo, Purist or McpCount}

\item{norm}{normlisation to apply if needed, can be none(by default), uq or vst}

\item{scale}{scale, can be raw, sc=sample center, gc=gene center, gsc=gene scale center, ssc=sample scale center (none by default)}

\item{newexp}{gene expression matrix/dataframe, sample in columns, gene in rows}
}
\value{
a data frame with row names as colnames of newexp, columns are different according the signatures, can be signature score or/and signature conclusion
}
\description{
Title applySignature
}
\examples{
applySignature(raw_counts,geneannot,signature="Gempred",norm="uq",scale="ssc")
}
