% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSignature.R
\name{callSignature}
\alias{callSignature}
\title{Title callSignature}
\usage{
callSignature(
  matrix,
  geneSymbols,
  signature = c("Gempred", "Puleo", "McpCount", "Purist"),
  toNorm = "none",
  toScale = "none"
)
}
\arguments{
\item{geneSymbols}{gene symbols, a vector of same length as the number of rows in newex}

\item{signature}{signature to apply, can be Gempred, Puleo, McpCount or Purist}

\item{toNorm}{normlisation to apply if needed, can be none(by default), uq or vst}

\item{toScale}{scale, can be raw, sc=sample center, gc=gene center, gsc=gene scale center, ssc=sample scale center (none by default)}

\item{newexp}{gene expression matrix/dataframe, sample in columns, gene in rows}
}
\value{
a data frame with row names as colnames of newexp, columns are different according the signatures, can be signature score or/and signature conclusion
}
\description{
Title callSignature
}
\examples{
callSignature(raw_counts,geneannot,signature="Gempred",toNorm="uq",scale="ssc")
}
