% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CancerRNASig_data}
\alias{CancerRNASig_data}
\alias{GP2model_simple}
\alias{molGradsys}
\alias{signatures}
\alias{t_GemPred_Biopsy}
\alias{.estimategenes}
\alias{.mcpgenes}
\alias{.puleoICAgw}
\title{Datasets included in CancerRNASig}
\format{
Various objects used internally by the package.

An object of class \code{data.frame} with 2120 rows and 3 columns.

An object of class \code{list} of length 4.

An object of class \code{list} of length 3.

An object of class \code{data.frame} with 2112 rows and 2 columns.

An object of class \code{data.frame} with 2 rows and 142 columns.

An object of class \code{data.frame} with 111 rows and 3 columns.

An object of class \code{data.frame} with 11895 rows and 10 columns.
}
\source{
Literature-derived gene signatures and models.
}
\usage{
GP2model_simple

molGradsys

signatures

t_GemPred_Biopsy

.estimategenes

.mcpgenes

.puleoICAgw
}
\description{
Collection of datasets used for RNA signature prediction and
tumor microenvironment characterization.
}
\keyword{datasets}
