% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{CancerRNASig_imports}
\alias{CancerRNASig_imports}
\title{Import required base R functions}
\description{
These imports make certain functions from stats and utils visible to R CMD check.
}
\keyword{internal}
