% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetSig.R
\name{coexpScore}
\alias{coexpScore}
\title{Title coexpScore}
\usage{
coexpScore(newexp, genesetlist)
}
\arguments{
\item{newexp}{A new expression dataset, should be gene symbol in rows}

\item{genesetlist}{Gene sets (signature or pathways) scores}
}
\value{
Gene set score
}
\description{
Title coexpScore
}
