% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimate.R
\name{.estimate}
\alias{.estimate}
\title{Title qEstimate}
\usage{
.estimate(ds, platform = c("illumina", "affymetrix", "agilent"))
}
\arguments{
\item{ds}{input dataset}

\item{platform}{platform type (e.g. "affymetrix", "agilent", "illumina")}
}
\description{
Title qEstimate
}
