% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GemPred_Biopsy.R
\name{GemPred_biopsy}
\alias{GemPred_biopsy}
\title{Title GemPred_Biopsy}
\usage{
GemPred_biopsy(newexp, gnt = "sc", pnt = "raw")
}
\arguments{
\item{newexp}{gene expression matrix/dataframe, sample in columns, gene in rows}
}
\value{
a data frame with row names as colnames of newexp, first column is gempred score
}
\description{
WARNING: This tool is intended for research purposes only. It is not intended for diagnostic use or clinical decision-making.
}
\examples{
# Example of how to call GemPred_Biopsy
# result <- CancerRNASig:::.GemPred_Biopsy(newexp = gene_expression_matrix)
}
\keyword{internal}
