% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{genesymUniqExp}
\alias{genesymUniqExp}
\title{Title genesymUniqExp}
\usage{
genesymUniqExp(newexp, geneSymbols, scoreFunc = NULL)
}
\arguments{
\item{newexp}{gene expression matrix/dataframe, sample in columns, gene in rows}

\item{geneSymbols}{gene symbols, a vector of same length as the number of rows in newex}

\item{scoreFunc}{a function taking a matrix and returning a score per line. this score is then
used to select the highest scoring row if several correspond to the same gene symbol.
With default rowSds, this means selecting the row with the highest variance.}
}
\value{
a data frame with row names as colnames of newexp, first column is purist subtype, second is purist score
}
\description{
This function aims to obtain a gene expression matrix with a unique gene expression value
for each provided gene symbol. By default, the gene
}
