% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McpCount.R
\name{MCPcounter}
\alias{MCPcounter}
\title{Title MCPcounter}
\usage{
MCPcounter(newexp, geneSymbols)
}
\arguments{
\item{newexp}{gene expression matrix/dataframe, sample in columns, gene in rows}

\item{geneSymbols}{gene symbols, a vector of same length as the number of rows in newex}
}
\value{
a data frame with samples in row and MCPcounter population quantification in columns
}
\description{
Title MCPcounter
}
