% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdacMolGrad.R
\name{projectMolGrad}
\alias{projectMolGrad}
\title{Title projectMolGrad}
\usage{
projectMolGrad(
  newexp,
  geneSymbols,
  normalize = c("newRef", "sameAsRef", "raw")
)
}
\arguments{
\item{newexp}{gene expression matrix or dataframe with gene in row(names) and samples in columns(names).}

\item{geneSymbols}{vector of gene symbols for the newexp dataset (simply set to rownames if the newexp is already in single values per gene symbols)}

\item{normalize}{Normalization (i.e. calibration) of the molecular gradient systems}
}
\value{
data frame of four projections based on the molecular gradients computed from four different types of expression datasets
}
\description{
Project a transcriptomic dataset on the Pancreatic Adenocarcinoma Molecular Gradient
Will throw an error if there is only one sample (one column)
}
\examples{
g <- rownames(CancerRNASig::molGradSys$PDX$gw)
projectMolGrad(matrix(rnorm(length(g) * 10), ncol = 10), g)
}
\keyword{internal}
