% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callSignature.R
\name{callSignature}
\alias{callSignature}
\title{Title callSignature}
\usage{
callSignature(
  matrix,
  geneSymbols,
  signature = NULL,
  normType = "raw",
  scaleType = "raw"
)
}
\arguments{
\item{matrix}{gene expression matrix/dataframe, sample in columns, gene in rows}

\item{geneSymbols}{gene symbols, a vector of same length as the number of rows in matrix}

\item{signature}{signature to apply, can be Gempred, tGempred, Puleo, pdacMolGrad, Mcpcount or Purist}

\item{normType}{normlisation to apply if needed, can be raw (by default), uq or vst}

\item{scaleType}{scale to apply if needed, can be raw (by default), sc=sample center, gc=gene center, gsc=gene scale center, ssc=sample scale center}
}
\value{
a data frame with row names as colnames of newexp, columns are different according the signatures, can be signature score or/and signature conclusion
}
\description{
Title callSignature
}
\examples{
callSignature(matrix,geneannot,signature="Gempred",normType="uq",scale="ssc")
}
