% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McpCount.R
\name{MCPcounter}
\alias{MCPcounter}
\title{Title MCPcounter}
\usage{
MCPcounter(genemat)
}
\arguments{
\item{genemat}{gene expression matrix/dataframe, sample in columns, gene in rows 
rownames should be gene symbols}
}
\value{
a data frame with samples in row and MCPcounter population quantification in columns
}
\description{
Title MCPcounter
}
